+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - **FEATURE:** Farben für light/dark-Mode können in den Einstellungen getrennt konfiguriert werden.
 - **FEATURE:** URL zum JDownloader kann in *Einstellungen/Erweitert* konfiguriert werden. Vielen Dank an @robNice.
 - **FEATURE:** Filme können an *pyLoad* zum Laden übergeben werden. Vielen Dank an @robNice.
+- **FEATURE:** Das Format der Zeitdarstellung im Tab Filme kann via *Einstellungen\Allgemein* konfiguriert werden.
 - interne Optimierungen für mehr Geschwindigkeit.
 - interne Fehler wurden behoben.